Mornin Arantxa-Jose,
We offer a 10% discount for stays longer than 28 days or longer.

This means if you were to book for a month, you would pay 12960 ZAR, instead of the usual 14400 ZAR.

If you would like to book for a month, please let me know which dates you would like to book for, then I can arrange that for you, as we normally allow maximum 21 days per stay.

Please note that the next available slot to book for a full month starts January 2023 till the end of April 2023.
The next 2 months (November and December we already have ongoing bookings).

Kind regards,
Lyall (Co-host)